site: move docs page bodies out of the route tree so One can read generateStaticParams
Team-Machine-Session: m12700
