+++ code/tamagui.dev/app/(site)/(docs)/docs/intro/[slug].tsx
@@ -1,88 +1,7 @@
-import { ThemeTint } from '@tamagui/logo'
-import { getMDXComponent } from '@vxrn/mdx-rust/client'
-import { useLoader } from 'one'
-import { useMemo } from 'react'
-import { HeadInfo } from '~/components/HeadInfo'
-import { SubTitle, nbspLastWord } from '~/components/SubTitle'
-import { TamaguiExamples } from '~/components/TamaguiExamples'
-import { DocsPageFrame } from '~/features/docs/DocsPageFrame'
-import { useDocsMenu } from '~/features/docs/useDocsMenu'
-import { components } from '~/features/mdx/MDXComponents'
-import { getOgUrl } from '~/features/site/getOgUrl'
-import { HomeH1 } from '~/features/site/home/HomeHeaders'
-
-export async function generateStaticParams() {
-  const { getAllFrontmatter } = await import('~/features/mdx/getMDXBySlug')
-  const frontmatters = getAllFrontmatter('data/docs/intro')
-  const paths = frontmatters.map(({ slug }) => ({
-    slug: slug.replace(/.*docs\/intro\//, ''),
-  }))
-  return paths
-}
-
-export async function loader(props) {
-  const { params, search } = props
-  const { getMDXBySlug, getCompilationExamples } =
-    await import('~/features/mdx/getMDXBySlug')
-  const { getDocsMode } = await import('~/features/docs/isTailwindMode')
-  const mode = getDocsMode(props)
-  const { frontmatter, code } = await getMDXBySlug(`data/docs/intro`, params.slug, {
-    mode,
-  })
-  return {
-    frontmatter,
-    code,
-    examples: getCompilationExamples(),
-    search,
-  }
-}
-
-export default function DocIntroPage() {
-  const { code, frontmatter, examples, search } = useLoader(loader)
-  const { next, previous } = useDocsMenu()
-
-  if (!frontmatter || !code) {
-    console.warn(`No frontmatter/code?`)
-    return null
-  }
-
-  const Component = useMemo(() => getMDXComponent(code), [code])
-
-  const GITHUB_URL = 'https://github.com'
-  const REPO_NAME = 'tamagui/tamagui'
-  const editUrl = `${GITHUB_URL}/${REPO_NAME}/edit/master/code/tamagui.dev/${frontmatter.slug}.mdx`
-
-  return (
-    <DocsPageFrame
-      headings={frontmatter.headings}
-      editUrl={editUrl}
-      next={next}
-      previous={previous}
-      frontmatter={frontmatter}
-      initialSearch={search}
-    >
-      <HeadInfo
-        title={`${frontmatter.title} — Tamagui`}
-        description={frontmatter.description ?? ''}
-        openGraph={{
-          images: [
-            {
-              url: getOgUrl({
-                title: frontmatter.title,
-                description: frontmatter.description ?? '',
-                category: 'intro',
-              }),
-            },
-          ],
-        }}
-      />
-      <HomeH1>{nbspLastWord(frontmatter.title)}</HomeH1>
-      <SubTitle>{nbspLastWord(frontmatter.description || '')}</SubTitle>
-      <ThemeTint>
-        <TamaguiExamples.Provider value={examples}>
-          <Component components={components as any} />
-        </TamaguiExamples.Provider>
-      </ThemeTint>
-    </DocsPageFrame>
-  )
-}
+// see the note in docs/core/[slug].tsx: the body lives outside the route tree so
+// this module stays a pure re-export and keeps its export names in the build
+export {
+  DocIntroPage as default,
+  loader,
+  generateStaticParams,
+} from '~/features/docs/pages/IntroDocPage'
