+++ code/tamagui.dev/app/(site)/(unstyled)/unstyled-ui/[...subpath].tsx
@@ -1,5 +1,7 @@
+// re-export the components docs page. the layout sets unstyled mode. the page is
+// exported as the named `DocComponentsPage`, so map it to default here
 export {
   DocComponentsPage as default,
   loader,
   generateStaticParams,
-} from '~/app/(site)/(docs)/ui/[...subpath]'
+} from '~/features/docs/pages/ComponentsDocPage'
