+++ code/tamagui.dev/features/docs/pages/GuidesDocPage.tsx
@@ -1,0 +1,77 @@
+import { ThemeTint } from '@tamagui/logo'
+import { getMDXComponent } from '@vxrn/mdx-rust/client'
+import { useMemo } from 'react'
+import { useLoader } from 'one'
+import { HeadInfo } from '~/components/HeadInfo'
+import { SubTitle, nbspLastWord } from '~/components/SubTitle'
+import { DocsPageFrame } from '~/features/docs/DocsPageFrame'
+import { useDocsMenu } from '~/features/docs/useDocsMenu'
+import { components } from '~/features/mdx/MDXComponents'
+import { getOgUrl } from '~/features/site/getOgUrl'
+import { HomeH1 } from '~/features/site/home/HomeHeaders'
+
+export async function generateStaticParams() {
+  const { getAllFrontmatter } = await import('~/features/mdx/getMDXBySlug')
+  const frontmatters = getAllFrontmatter('data/docs/guides')
+  const paths = frontmatters.map(({ slug }) => ({
+    slug: slug.replace(/.*docs\/guides\//, ''),
+  }))
+  return paths
+}
+
+export async function loader(props) {
+  const { params, search } = props
+  const { getMDXBySlug } = await import('~/features/mdx/getMDXBySlug')
+  const { getDocsMode } = await import('~/features/docs/isTailwindMode')
+  const mode = getDocsMode(props)
+  const { frontmatter, code } = await getMDXBySlug('data/docs/guides', params.slug, {
+    mode,
+  })
+  return {
+    frontmatter,
+    code,
+    search,
+  }
+}
+
+export function DocGuidesPage() {
+  const { code, frontmatter, search } = useLoader(loader)
+  const { next, previous } = useDocsMenu()
+  const Component = useMemo(() => getMDXComponent(code), [code])
+
+  const GITHUB_URL = 'https://github.com'
+  const REPO_NAME = 'tamagui/tamagui'
+  const editUrl = `${GITHUB_URL}/${REPO_NAME}/edit/master/code/tamagui.dev/${frontmatter.slug}.mdx`
+
+  return (
+    <DocsPageFrame
+      headings={frontmatter.headings}
+      editUrl={editUrl}
+      next={next}
+      previous={previous}
+      frontmatter={frontmatter}
+      initialSearch={search}
+    >
+      <HeadInfo
+        title={frontmatter.title}
+        description={frontmatter.description || ''}
+        openGraph={{
+          images: [
+            {
+              url: getOgUrl({
+                title: frontmatter.title,
+                description: frontmatter.description ?? '',
+                category: 'intro',
+              }),
+            },
+          ],
+        }}
+      />
+      <HomeH1>{nbspLastWord(frontmatter.title)}</HomeH1>
+      <SubTitle>{nbspLastWord(frontmatter.description || '')}</SubTitle>
+      <ThemeTint>
+        <Component components={components as any} />
+      </ThemeTint>
+    </DocsPageFrame>
+  )
+}
